requirements.txt
albumentations==1.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
docopt==0.6.2
Flask==2.1.2
greenlet==1.1.2
idna==3.3
imageio==2.19.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
networkx==2.8
numpy==1.22.3
opencv-python-headless==4.5.5.64
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pipreqs==0.4.11
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.27.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
SQLAlchemy==1.4.36
threadpoolctl==3.1.0
tifffile==2022.5.4
timm==0.5.4
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
yarg==0.1.9
